save task info changes before switching branches
