+++ app.json
@@ -12,7 +12,7 @@
       "infoPlist": {
         "NSLocationWhenInUseUsageDescription": "This app uses your location to show your position on the map.",
         "NSPhotoLibraryUsageDescription": "This app needs access to your photo library to update your profile picture."
-    }
+      }
     },
     "android": {
       "adaptiveIcon": {
@@ -33,7 +33,8 @@
           "resizeMode": "contain",
           "backgroundColor": "#ffffff"
         }
-      ]
+      ],
+      "expo-router"
     ]
   }
 }
