+++ package.json
@@ -15,16 +15,16 @@
     "@react-navigation/bottom-tabs": "^7.3.10",
     "@react-navigation/elements": "^2.5.2",
     "@react-navigation/native": "^7.1.6",
-    "expo": "^53.0.13",
+    "expo": "53.0.15",
     "expo-blur": "~14.1.5",
     "expo-constants": "~17.1.6",
     "expo-font": "~13.3.1",
     "expo-haptics": "~14.1.4",
-    "expo-image": "~2.3.0",
+    "expo-image": "~2.3.1",
     "expo-image-picker": "~16.1.4",
-    "expo-linking": "~7.1.5",
-    "expo-location": "~18.1.5",
-    "expo-router": "~5.1.1",
+    "expo-linking": "~7.1.6",
+    "expo-location": "~18.1.6",
+    "expo-router": "~5.1.2",
     "expo-splash-screen": "~0.30.9",
     "expo-status-bar": "~2.2.3",
     "expo-symbols": "~0.4.5",
